GeanyGenDoc: Add policy PASS to completely ignore a symbol
This is useful e.g. to fix some annoyances of some tag parser that
extracts for example all the variables even inside functions (such as
the PHP one), which clutters the symbol list and then makes many
documentation requests to fail simply because they matched a variable
and not the enclosing function.

git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@1592 e8383189-8249-0410-b506-acc975200cbd