GeanyGenDoc: Don't copy system file to when opening filetype configuration
Don't unconditionally copy the system configuration file to the user
configuration directory when user calls "Edit Current Language Configuration".
Now we rather open the current file (as usually chosen) without copying it, so
it gets written to the user configuration directory upon save but not before.

This avoids copying useless configuration files to the user directory that may
hide new versions of the system configuration.

git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@1789 e8383189-8249-0410-b506-acc975200cbd