GeanyGenDoc: Fix file names encoding usage
Now the file names should be correctly converted from an to the correct
encoding. This may improve support of systems where the encoding is not
UTF-8 (Windows, a small portion of Unices, ...).

Note that this code was not tested under non-UTF-8 systems, especially
not with paths that may cause problems. Further testing would be
needed, let's wait for bug reports :)

git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@1794 e8383189-8249-0410-b506-acc975200cbd