WebHelper: Avoid useless notifications when synchronizing settings from widgets
When synchronizing settings from their display widget, only change the
corresponding property if the value actually changed. This avoids
useless notifications if the value was not modified.

This is particularly useful because we synchronize on preference dialog
validation, but since the dialog is shared by all plugins we might
synchronize even when the user has only changed the settings of another
plugin.

git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@1871 e8383189-8249-0410-b506-acc975200cbd