TreeBrowser: Fix crash and memory leak in treebrowser.c.
In path_is_in_dir(), if it's about to return NULL (if the function parameters
have no path components in common) it tried to free() a string
literal, which causes a crash. In the same function, there was a memory leak,
because diffed_path was reassigned without being free()'d before.

Fix both problems.

Patch by Thomas Martitz, thanks.

git-svn-id: https://geany-plugins.svn.sourceforge.net/svnroot/geany-plugins/trunk/geany-plugins@2065 e8383189-8249-0410-b506-acc975200cbd