markdown: Add code to Discount to return the internal buffer
This avoids the caller having to duplicate the internal string and
works for the simple case of the Markdown plugin.

Adds function `mkd_compile_document()` which simply compiles a const
string to a newly-allocated string, hiding the little bit of extra
code needed normally to do this.

There's tons of noise in the commit because I had trailing whitespace
stripping enabled and I re-indented some of the code to make it more
readable (to me).