pairtaghighliter: Fix crash on plugin unload
Fix crash on plugin unload if the plugin didn't run on any editor, or
if the last editor it ran on became invalid.

The global 'sci' variable wasn't set in 'plugin_cleanup()' and thus may
either have been uninitialized or initialized to a value not valid
anymore.

The simplest way to reproduce the crash was to load and then
immediately unload the plugin, without leaving the plugin manager
dialog.