pairtaghighliter: Remove a global variable
Remove use of the 'sci' global variable and instead pass it as an
argument to each function requiring this data.

The use of global variables makes code harder to understand and debug,
and often leads to unexpected issues.  Here, it lead to a crash when
unloading the the plugin because an apparently innocent call used this
global variable before its initialization.  This was fixed in previous
commit, but removing use of the global variable hopefully will prevent
such bugs from reappearing later.