pohelper: Fix infinite loop retrieving a message's translation
Fix an infinite loop in get_msgstr_text_at() if there was only default
style after it until the end of the document.

sci_get_style_at() returns 0 if the position was invalid, which happens
to be the value of SCE_PO_DEFAULT, so simply looping while it matched
entered an infinite loop at the end of the document.  So, additionally
make sure the position is inside the document's range.