Fix Warnings
Alias warnings by setting -fno-strict-aliasing flag, this disables
the optimisation so is safe (and matches Python).

String const warnings by -Wno-write-string flag, gcc notes that
the warning may be useful, but can be annoying when const is not
properly used on the pointers (defined by Python so not fixable).

geanypy-uiutils.c combo box warning by updating to GtkComboBoxText
to match changes to Geany API.