Autotools: Do not make a positive decision from GP_CHECK_PLUGIN_DEPS()
GP_CHECK_PLUGIN_DEPS() must never force-enable a plugin by setting
enable_$plugin=yes, even if the plugin was in auto mode and the
dependencies are met;  otherwise a further check would assume the
user forcefully enabled the plugin, and then would abort if it failed.

So, only disable the plugin or error out, but never enable it.  This
allows e.g. successive calls with one failing:

	GP_CHECK_PLUGIN_DEPS(myplugin, A, existing-package)
	GP_CHECK_PLUGIN_DEPS(myplugin, B, missing-package)

This should disable plugin "myplugin" by default;  but it would have
had aborted on the second check because the first one would have set
"enable_myplugin" to "yes" instead of "auto".