spellcheck: fix checking with cppcheck
cppcheck tests all preprocessor conditionals, but when the code tests
whether one is defined or not it tests simply with it defined or not,
but always to an empty value.  So, code testing whether a macro is
defined, but then relying on the macro expanding to a non-empty value,
cppcheck might very well detect a syntax error.

So, explicitly tell cppcheck to only test the SCE_PAS_DEFAULT=0 case
so it doesn't incorrectly test the SCE_PAS_DEFAULT= case.