geanylatex: fix automatic capitalization of multi-byte characters
Fix indexing relative document positions by characters, not bytes, when
performing automatic capitalization of sentence start.

Geany API dependency is bumped to 217.  Even though this change doesn't
actually require new API introduced by this version, it requires
Scintilla 3.3.5 which is only available in Geany 1.24, and API 217 is
the closest one to this.