spellcheck: Don't check default styles of embedded HTML scripts
Don't check the spelling for the default styles used by embedded
scripting languages inside HTML.

While spellchecking the default HTML and SGML styles makes sense since
it includes most text, it doesn't make sense to do so for default style
of embedded scripting languages (such as PHP, Python, etc.), as those
"default" styles are the bare identifiers of the language (classes,
functions, etc.).

This doesn't affect HTML or SGML embedded in scripting languages, as
the default style for the scripting languages have a different value
than the ones for the markup languages.