commander: Sorting: give higher score for basename matches
Give higher scores for matches inside the last path component, as it is
more likely the user wanted to match it.

For example it helps matching against filenames that share prefix with
another path component, as it would otherwise also match other files
under the same path, making it harder to find the correct file.