autotools: Don't allow undefined symbols in plugins when possible
Tell the linker not to allow undefined symbols in built plugins it
possible.

This avoids use of non-plugin API functions from plugins as using those
would lead to undefined symbols.  This works even as we don't link to a
"libgeany" or the like because Geany plugin API is hidden behind macros
indirection.