multiterm: Fix building with recent Vala versions
Vala doesn't allow dynamic code as global variable initializers
anymore, so move the dynamic initialization inside plugin_init().

Interestingly, if this actually used to work is only because the
initialization wasn't actually dynamic in C, as a NULL GList is a
valid one and that's what it was initialized to.