multiterm: Fix vapidir under Automake >= 1.11.0a with VPATH builds
Automake 1.11.0a introduced an incompatible change, that is it now
changes directory to `$(srcdir)` before running `valac`.

This means `$(top_srcdir)`-based paths are most likely to be incorrect
when building out-of-tree (VPATH) if the user used relative path as
srcdir (e.g. invoked configure with something like `../configure`).

To fix that, use the absolute version of `$(top_srcdir)`.