geanyctags: open document automatically only when there is a single result
If there are several results from the ctags search, opening the first document
isn't the best thing to do because it's most probably not the tag we are
searching for. Instead, don't open the document and just write all the
results to the messages window so the user can choose which one to open.