debugger: Make GCC happy
Apparently GCC is not smart enough to understand that when the a NULL
name is passed to `gdb_mi_result_foreach_matched` the `strcmp()` call
cannot possibly happen.  Work around the warnings it emits by adding an
explicit (and useless) check on the `strcmp()` argument, which probably
will be optimized out anyway.