debugger: Use the GDB/MI module to parse GDB output in the GDB module
This is likely to *NOT* be completely compatible with what the original
code did.  There are some explicitly different things that may either
have been errors in the old code or changes in GDB output.

The changes seem (basic testing) to work fine on Debian's GDB 7.7.1.

NOTE: There is currently no conversion from the locale encoding to
UTF-8.  The GDB/MI module decodes the 7bit strings from GDB, but they
probably encode locale-encoded data.  This means this probably won't
work reliably on non-UTF-8 locales without an appropriate conversion.
"Appropriate conversion" yet has to be defined -- but as UTF-8 covers
all of Unicode it probably is OK to convert back and forth even when
not strictly required, it should at worse be useless.