git-ui: Properly change which ref is monitored when HEAD changes
This fixes ref monitoring after switching branches.  While in most
cases it doesn't matter because either the file has not changed or it
has to be reloaded from disk anyway, it happens when the user uses
`checkout -b`, as this will change the ref HEAD points to without
touching the tree or index.