git-changebar: Fix writing of color values in the configuration
Incorrect variable reference in `sizeof` expression lead to using the
size of a pointer as the length of the buffer.  This used to work on
64 bits systems because the buffer is 8 bytes in length, which happens
to be the size of a pointer on 64 bits systems, but it fails on 32 bits
ones because the size of the pointer is 4 bytes there, which is too
small for what we actually want to write, leading to a truncation of
the value.

Thanks to Mark Burton for discovering the bug and providing a patch!