GeanyPy: Unload Python plugins when GeanyPy is deactivated
This prevents crashes and/or lockups when the GeanyPy C plugin
is reloaded.

FIXME: Because it calls the normal unloading path, it updates
the list of loaded plugins to remove each plugin. The net effect
is that when the C GeanyPy plugin is unloaded, it forgets all
of the Python plugins that were loaded at the time.