projectorganizer: Load project even if sidebar isn't realized
Project organizer's sidebar tab isn't realized also when the initial
sidebar tab differs from the projectorganizer tab in which case the tree
isn't created until we switch to the projectorganizer tab.

This in general isn't a problem but some functions like "find project
file" rely on the existence of the tree model and don't work without
it (which happens before the first switch to the project organizer tab).