ProjectOrganizer: Fix too aggressive symlink cycle detection
The previous symlink cycle detection would check if the child directory is
within parent. This however breaks the case where in the root of the
project there are symlinks to directories outside the project directory.

Fix this by simply recording the resolved real paths during the tree
creation. Skip subtree creation if the path was visited previously.