lineoperations: fixed last newline, empty doc problems
Added `ensure_final_newline()` to add a newline if it does not exist
before operation. If the final line does not have an newline, and it
shifts position, the resulting line will be concatenation of two lines.

Added null terminating character to new_file string in case of empty
file.