debugger/gdb_mi: Do not allocate values when reading invalid input
Only allocate a value structure in parse_value() if the input actually
contains a readable value, instead of destroying it if it doesn't.

This is particularly useful as the implementation of parse_value()
itself recurses in order to discriminate a value from a result, which
means that a value would be allocated and freed immediately for each
result in the input.

This gives a speedup of about 10% on typical data.