debugger: Greatly improve Stack Tree population speed
Appending elements to a TreeStore is very slow, while prepending them
is fast.  As we add all items at once, prepend elements in reverse
order, which is a lot faster and does the same.

This leads to tremendous speedups with large stack traces.
For example, showing a 105000 frames height trace (resulting from
infinite recursion and thus stack exhaustion) used to take more than
105s, while it now takes less than 0.3s.