debugger: stree: Don't maintain a table of thread row references
GtkTreeRowReferences are not free to maintain as they might get updated
whenever a row is inserted or removed, and it's unlikely there would be
a very large amount of threads requiring great scaling.

So drop the reference logic for a simple search in the existing threads
whenever we need to find one out, which doesn't happen that often
anyways.