debugger: Fix marker for current frame
Display the correct marker for the current frame when there is several
frames at the same location (e.g. from recursion).

As Scintilla markers are drawn in their ID order, the marker one wants
to see "on top" should have the highest ID.  So, make the current
frame marker (M_CI_ARROW) have a higher ID than the generic frame one
(M_FRAME) so if both are set on the same line the current frame one is
the one on top (e.g. visible, as they have the exact same shape).