Geniuspaste: Install a custom URL handler for Paste link
This workarounds a strange crash on Windows when the default GTK
URL handler is called.
We need the URL handler only on the secondary text label but we
cannot access it directly, so iterate over all labels of the message
dialog's message area and attach the custom handler to all of the
labels. This shouldn't harm much as only the secondary label contains
a link.