projectorganizer: Fix multiple tabs in Project Properties
There is a problem when there are multiple tabs added by plugins: if
on_project_dialog_open() is called and Project Organizer stores
the index of the tab and then another plugin adds a tab before the
Project Organizer's tab, then the stored index becomes invalid which
can lead to not correctly removing the tab from the dialog.

Instead store a pointer of the added tab widget and remove the widget
when the properties dialog is closed.