debugger: Fix escaping target executable path if it contains non-ASCII
GDB doesn't like octal escape sequences in executable path passed to
`-file-exec-and-symbols`, yet it accepts them (but doesn't require
them) in arguments to `-break-insert`.

So, use an escaping valid everywhere, as for the moment it's unclear
whether there is any benefit to escaping special bytes in argument to
`-break-insert`, and so consistency is better.

Fixes #416.