Addons: Implement unselect with single click
For this to work, we need to connect to the "button-press-event" for
each document, then process each single click and delay the detected
double click event by 50ms before processing it.
This way all markers will be removed (in case of double click
actually twice) and new markers will be set on double click.

Due to the new behaviour introduced here, the feature is disabled
by default.

Closes #445.