TreeBrowser: more mnemonics in the context menu
This menu has many useful items (I use at least 6 of them daily),
but finding them manually with Up/Down or the mouse is annoying.

The choice of mnemonics is somewhat arbitrary, but they are unique.

Note that TreeBrowser has keybindings for a few of these operations,
but they are not quite the same thing, because keybindings are
global to Geany, and the global keybinding space is sort of crowded
for me already (hard to pick a convenient but unused one).