TreeBrowser: pop up context menu on Shift+F10
GTK+ by default maps Shift+F10 to the same thing as Menu,
which is useful for laptop keyboards (that have no Menu key).
But because TreeBrowser handles keypresses on its own,
it doesn't get this functionality automatically.

Implementation is per the "Test for modifier keys correctly"
section of the GTK+ reference manual.