git-changebar: correctly display diff when first line removed
Diffs for removed lines are shown on previous line - this works well except
the first line where there's no previous line so nothing is displayed.

This patch adds corresponding tests to the code and shows diffs for the
first line on the first line. It also fixes related problems with going
to previous/next hunk and undoing hunks.