Improve the error message on attempting to run gdb
On failure to run gdb, Scope displays only the generic error message,
such as "The system cannot find the file specified", giving the
impression that the executable to be debugged is missing or has some
other problem. This patch changes the message to "<gdb name>: <error>",
for example "gdb: The system cannot find the file specified" (unlikely
under *nix, but gdb may be missing from the PATH under Windows).