git-changebar: Move cursor at the end of the undone hunk text for deleted hunk
For deleted hunk, move cursor at the end of the undone text. This
behavior corresponds to what normal undo does - when undoing deletion,
cursor is also placed at the end of the newly inserted text.

Also change primary position for SCI_SCROLLRANGE in this case so it becomes
the cursor position.