Initialize loop variables to make cppcheck happy
It seems newer cppcheck versions are too sensible to uninitialized loop
variables when the loop is created by a macro.
This might be a bug in cppcheck or it just cannot check the loop
properly. By initializing the affected variables explicitly, cppcheck
is happy while the unnecessary initialization should not have any
impacts.