projectorganizer: Use stock icons instead of the ones provided by the plugin
I used to ship toolbar icons with the plugin to be sure they are correct semantically.
For instance, the "+" icon is used to expand the tree and the semantics of the
corresponding icon in GTK is "add" - some crazy theme could for instance use
word "ADD" in the icon and then the icon wouldn't correspond to the "expand"
meaning.

However, I've just been playing with many icon themes on OS X and they display
these basic icon in more or less the same way so using stock icons seems to be
relatively safe. This also solves the problem of the shipped icon being displayed
in a different way than the rest of the system icons and in addition, if HiDPI icons
are available, they can be used on a highres monitor.