workbench: import Geany project data
The workbench plugin now imports the "base_path" and "file_patterns" from Geany's project settings
as an workbench project directory. That directory is then treated as the project's base dir.

In the config file all settings related to the project's base dir are prefixed with "Prj-"
instead of e.g. "Dir1-". On loading of a workbench's project, the plugin checks if there
is a "base_path" in the "project" section but no "Prj-BaseDir" in the "Workbench" section.
In that case the data from the Geany project is imported. That means, the import happens only once.
On import, the project is marked as changed (a '*' is displayed after the project's name).

The project's base dir is always labeled "Base dir" and has a specific icon to show that it differ's
from the other directories. Also, it is always listed first/on top of the tree.