addons: show color tip and start Color Chooser with double click
This new feature shows a color tip if the mouse is hovered over a
color value like e.g. '#fff' or '#ffffff'. If a color value is double
clicked, then the Color Chooser is started. Both (showing the color tip
and starting the Color Chooser) can be enabled and disabled via the
plugin preferences. See #663.