workbench: remove files or directories.
The user can now remove a selected file or directory. Closes #711.
The user has to confirm the action and is warned that it cannot be undone.

In case a directory is removed, then first all files will be removed.
The workbench plugin only removes that files in a directory which are
displayed in the sidebar. That means files that are not displayed because
they do not match the file patterns will not be removed. After removing
all files, all sub-directories and finally the selected directory will
be removed. This will fail if a directory is not really empty.

If removing of a file or directory fails, then a message is output.