workbench: fix broken live-update
This fixes the broken live-update. The feature did not work since commit
dcc37957e917a2e8c2f3e5d483c9a832eceff1e0. The reason was that the sub-directories
were not added to the file-table and so no file monitor was created for them.
The live-update only worked for the main directories but not for their sub-directories.