scope: Don't place AM_CONDITIONAL inside a shell condition
This is not valid, because the AM_CONDITIONAL has to be known by
Automake in all cases.  Automake is smart enough to detect it if used
in a AutoShell conditional (like AS_IF/AS_CASE/...), but can't work any
magic inside a pure shell conditional (if/case/...).

So, move the AM_CONDITIONAL outside the conditional, as the value is
already conditional anyway.