TreeBrowser: Always search the whole path when trying to find current file in tree
If the current file path is not part of the current directory,
the code starts at / to find the path. However, it always skipped the
first path part (i.e. /) and so never matched.
This happened most often when Geany's default open directory has
no common parts with the file path of the current document (e.g.
Geany's default open directory is /tmp and the file path is like
/home/user/src/somefile.c).