scope: Improved evaluation on-hover
Scope automatically evaluates an expression if the mouse pointer is hovered over it.
But before this change it only evaluated words. E.g. if the mouse was hovered over
"structa->item1" then scope would have evaluated "structa" or "item1" depending on
the exact position of the mouse pointer. With this change it will instead evaluate
the value of "structa" if the mouse pointer is over "structa". If the mouse pointer
is over "item1" then the expression "structa->item1" will be evaluated.