workbench: added new feature "Search projects"
The feature lets the user search for projects to add them to the Workbench.
After selecting the menu item, the user can select a directory to search through.
The directory and it's sub-directories are scanned for project files (".geany").
All found files are listed and after scanning is done the user can select the
projects which shall be added to the Workbench.

The dialog is not modal - the user can do other work during the scanning of the
directory.