workbench: added option to let git decide which files to display
In the directory settings it is now possible to choose between a
workbench or a git file filter. The later is new and lets git decide
which files to display depending on the contents of the .gitignore
file.

This also adds a new function to the utils-lib called
'gp_filelist_scan_directory_callback()'.