pairtaghighlighter: do not stop scanning on linebreaks
Before this change the plugin stopped scanning for a tag on the start or the
end of a line. Therefore tags which were split over multiple lines were not
recognized. With this change the plugin scans forward or backward up to a
certain limit and does not stop on a linebreak. The default limit is 1024 bytes
but it can be changed by configuration if required. Fixes #775.