treebrowser: Use better icons
For expanded directory we should use the "folder-open" icon and not
"document-open" which is an action icon used for e.g. a toolbar icon
to open a document and can be themed in a completely different way.

Similarly, the "folder" icon used in the toolbar is a document-type icon
and here we should use some action icon. I used the "edit-copy" icon
which typically looks like two documents next to each other which kind
of corresponds to this action. But it's possible there's some better
option here.