projectorganizer: Make showing empty directories in the sidebar configurable
When file operations were added to project organizer, the plugin started
showing empty directories in the sidebar because they are useful e.g.
for creating new files inside them. This may be unwanted e.g. in projects
with many directories which are irrelevant to the code.

This patch makes this configurable on a per-project basis.

Fixes #881