Update README - "Mini Script"
"gms" is an abbreviation used in this document - "gms" is not a kind of Mini Script; it ***IS*** Mini Script...

Used uppercase for the abbreviation as it is not the name of an executable. And properly introduced the abbreviation.

Used the name "Mini Script" as that is what it is called in the Plugin Manager - though it is currently not consistently spelled in the software.

Dropped the word "configuring" as there isn't any one-time configuration part - only the options/settings in the Mini Script window.

General copy editing (e.g. ref. <https://en.wikipedia.org/wiki/Perl>, <https://en.wikipedia.org/wiki/Python_%28programming_language%29>, and <https://en.wikipedia.org/wiki/AWK>).