Markdown: Update using known GeanyDocument when available
For whatever reason, in some cases `document_get_current()` doesn't
return a valid document when it seems like it should, so when updating
the markdown preview from signals where the related GeanyDocument
is available, use that instead of calling `document_get_current()`.

In other cases, continue to use `document_get_current()` as before.

Closes #1062