Windows: Support specifying installation target for silent installation
The installation directory for Geany-Plugins is automatically determined
by inspecting the existing Geany installation to use the same directory.
But it might be desired for silent installations (i.e. without user
interaction) to specify an alternative installation directory by
appending "/D=c.\some-dir" to the command line. This is now possible.

Used for CI builds to test the installation automatically.