Project Organizer: Set header filetype to match source filetype
When opened, the filetype of header files is changed to match
the source filetype.  This is particularly useful for *.h files,
which may be C or C++.

Out of project files are matched with files in the same directory.
Common code between swap header/source and change header filetypes
are combined.  When a project is not loaded, default file extensions
are used.