vimode: avoid caret position change when inserting EOL at EOF
When current position is changed in the document and this position is
outside the visibility of the current screen, when returning
back, the call of SCI_GOTOPOS causes window scroll to a different position
than before.

To avoid this problem, insert_eof_nl_if_missing() and remove_char_from_eof()
avoid setting the position now and, instead, insert/remove the EOL
character using operations that don't change caret position.