vimode: fix escape not working when numlock is on
It seems that GDK_MOD2_MASK isn't the right modifier to check for
command pressed - its mapping is platform-specific and while this works
on macOS, it has undesirable side-effects on linux. Use GDK_META_MASK
instead which based on my testing seems to do the right thing both on
macOS and linux.