Project Organizer: Add ability to open file manager and terminal
File manager/terminal are opened at the location of the selected file
in the tree view.  If nothing is selected, the location of the current
document is used.  If the current document does not have a path,
the project directory is used.  If the project directory cannot be
located, the current directory is used.

The file manager is opened with the path as the first argument.

The terminal path is set using the PWD.