vimode: Don't build the viw binary on Windows
With the latest version of msys, the viw binary fails to link (most
probably it should be built as a C++ binary instead of the current plain
C binary).

Since this binary is used just for testing purposes, there's
no need to build it on Windows so this patch skips the build there to
avoid this problem.