git-changebar: Re-set first visible line after Scintilla size request
After updating to Scintilla 5.3.7 the Scintilla popup with the diff
against git shows the wrong number.

It seems that gtk_widget_set_size_request() scrolls Scintilla view so
despite it was set previously, it isn't set correctly after this call
and has to be re-set.

Note that the call to SCI_SETFIRSTVISIBLELINE has to stay at the original
location too, otherwise subsequent calculation using SCI_POINTXFROMPOSITION
doesn't work because the line may not be visible.