Autotools: regenerate gmo files when po files are updated
Make sure the gmo files are regenerated when the corresponding po files
are updated, not to have stale translations after updating a po file.

I had a hard time finding a rationale for the behavior of not
automatically updating the gmo files when their source changed; all I
could find is [1] which suggests it's intentional.  Still, at least
with our configuration of PO_DEPENDS_ON_POT=no, which means po files
are not updated all the time, the point seems moot and does more harm
than good.

Fix this by manually adding the correct dependency to the main target
so the gmo files are updated automatically when the po changes.
The only case where I can see it performing "useless" work is when
calling update-po, which will effectively update all po files, and thus
rebuild all gmo files.  But that doesn't sound like any kind of
problem, and rather expected.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=400453#24