geanygendoc: Add support for local variables
Since Geany 2.0 a lot of those are reported for several languages,
including C.  Those should usually not be documented, but their parent
should be instead, otherwise they "get in the way" of triggering
documentation of their parent.

Add support for the "local" match rule for local variables, and update
the default rules to use the FORWARD policy for them.  It's not
actually the parent documenting the child, but the effect is the same
as the user probably wanted to document the parent rather than the
local.