webhelper: Update minimal requirements
GTK, GLib, GIO, Gdk-Pixbuf, WebKit, all those are actually minimal
dependencies of the code (although arguably GIO is always at the same
version than GLib, so having something different there is somewhat of a
lie).  Geany is merely based on the PLUGIN_VERSION_CHECK(), which might
or might not be appropriate.