codenav/goto_file: Simplify reference model handling
The way the completion model to use when no directory separator is found
is unecessarily convoluted. This change simplifies and improves it.

Also fixes a bug that can cause junky behaviour when pasting a path as soon
as the dialog opens, if that was also the last path in the dialog the last
time it was opened. The root cause is that the completion model is set during
init to the reference directory, but directory_check() does not create a new
one for the pasted path, because curr_dir (which is static) has not been
updated to reflect that the reference mode is now active.