Delete indicators when a selection exists
When a user double clicked a word, the word was selected and highlighted with
automark, but no new highlightings appeared on scrolling the file. The word was
highlighted near the selection only.

In this case, it is better not to highlight this word anywhere, so as not to
confuse the user.