Avoid mnemonic conflicts
Avoid having two identical mnemonics in any menu when any plugin is
loaded, but for the Tools menu itself for now.

The downside is that for LSP the entries in the editor context menu do
not all have the same mnemonic as the equivalent one in the plugin's
tool menu anymore.  However, there currently isn't a better solution I
could find given some strings are also used in other menus by the
Project Organizer plugin, with a single translation for both.

The Tools menu isn't perfect either yet: not only having so many
entries when all plugins are loaded makes this hard if not impossible,
but there's also a few conflicts with Geany's own items that should be
avoided, but for which a fix is non-trivial and will require some more
thought.