Build with GDK_DISABLE_DEPRECATION_WARNINGS
Some of the GTK deprecation warnings are hard to fix and make the build
output too noisy. This is not good for development as it makes it
easy to overlook actually valuable warnings and in case of compilation
errors, it makes it hard to locate the error messages among all the
noise.

GDK_DISABLE_DEPRECATION_WARNINGS is used for compiling Geany too
so I think it would be valuable for g-p as well.